split helper functions and macros for about-facilities
