+++ gef-ui/server/controllers/confirm-cover-start-date/index.js
@@ -2,10 +2,9 @@ const { format } = require('date-fns');
 const
   {
     validationErrorHandler,
-    pastDate,
-    futureDateInRange,
     getEpoch,
   } = require('../../utils/helpers');
+const { pastDate, futureDateInRange } = require('../../utils/facility-helpers');
 const { applicationDetails } = require('../application-details');
 const api = require('../../services/api');
 const CONSTANTS = require('../../constants');
