+++ gef-ui/server/controllers/submit-to-ukef/index.test.js
@@ -2,7 +2,7 @@ import { submitToUkef, createSubmissionToUkef } from './index';
 import api from '../../services/api';
 import CONSTANTS from '../../constants';
 
-const { isNotice } = require('../../utils/helpers');
+const { isNotice } = require('../../utils/deal-helpers');
 
 jest.mock('../../services/api');
