+++ gef-ui/server/utils/MOCKS/MOCK_APPLICATIONS.js
@@ -1,6 +1,5 @@
 const CONSTANTS = require('../../constants');
 
-// TODO: DTFS2-5228 split file up
 const MOCK_AIN_APPLICATION = {
   _id: '61a7710b2ae62b0013dae687',
   dealType: 'GEF',
@@ -351,130 +350,6 @@ const MOCK_AIN_APPLICATION_ISSUED_ONLY = {
   },
 };
 
-const MOCK_UNISSUED_FACILITY = {
-  status: CONSTANTS.DEAL_STATUS.COMPLETED,
-  details: {
-    _id: '61a771cc2ae62b0013dae68a',
-    dealId: '61a7710b2ae62b0013dae687',
-    type: 'CASH',
-    hasBeenIssued: false,
-    name: 'Facility two',
-    shouldCoverStartOnSubmission: true,
-    coverStartDate: 1638403200000,
-    coverEndDate: '2030-01-01T00:00:00.000Z',
-    monthsOfCover: null,
-    details: [],
-    detailsOther: '',
-    currency: 'GBP',
-    value: 2000,
-    coverPercentage: 80,
-    interestPercentage: 1,
-    paymentType: 'IN_ADVANCE_MONTHLY',
-    createdAt: 1638363596947,
-    updatedAt: 1638442632540,
-    ukefExposure: 1600,
-    guaranteeFee: 0.9,
-    submittedAsIssuedDate: '1638363717231',
-    ukefFacilityId: '0030113305',
-    feeType: 'in advance',
-    feeFrequency: 'Monthly',
-    dayCountBasis: 365,
-    coverDateConfirmed: false,
-  },
-  validation: { required: [] },
-};
-
-const MOCK_ISSUED_FACILITY = {
-  status: CONSTANTS.DEAL_STATUS.COMPLETED,
-  details: {
-    _id: '61a771cc2ae62b0013dae68a',
-    dealId: '61a7710b2ae62b0013dae687',
-    type: 'CASH',
-    hasBeenIssued: true,
-    name: 'Facility one',
-    shouldCoverStartOnSubmission: true,
-    coverStartDate: 1638403200000,
-    coverEndDate: '2030-01-01T00:00:00.000Z',
-    monthsOfCover: null,
-    details: [],
-    detailsOther: '',
-    currency: 'GBP',
-    value: 2000,
-    coverPercentage: 80,
-    interestPercentage: 1,
-    paymentType: 'IN_ADVANCE_MONTHLY',
-    createdAt: 1638363596947,
-    updatedAt: 1638442632540,
-    ukefExposure: 1600,
-    guaranteeFee: 0.9,
-    submittedAsIssuedDate: '1638363717231',
-    ukefFacilityId: '0030113305',
-    feeType: 'in advance',
-    feeFrequency: 'Monthly',
-    dayCountBasis: 365,
-    coverDateConfirmed: false,
-    changedToIssued: true,
-  },
-  validation: { required: [] },
-};
-
-const MOCK_ISSUED_FACILITY_UNCHANGED = {
-  status: CONSTANTS.DEAL_STATUS.COMPLETED,
-  details: {
-    _id: '61a771cc2ae62b0013dae68a',
-    dealId: '61a7710b2ae62b0013dae687',
-    type: 'CASH',
-    hasBeenIssued: true,
-    name: 'Facility one',
-    shouldCoverStartOnSubmission: true,
-    coverStartDate: 1638403200000,
-    coverEndDate: '2030-01-01T00:00:00.000Z',
-    monthsOfCover: null,
-    details: [],
-    detailsOther: '',
-    currency: 'GBP',
-    value: 2000,
-    coverPercentage: 80,
-    interestPercentage: 1,
-    paymentType: 'IN_ADVANCE_MONTHLY',
-    createdAt: 1638363596947,
-    updatedAt: 1638442632540,
-    ukefExposure: 1600,
-    guaranteeFee: 0.9,
-    submittedAsIssuedDate: '1638363717231',
-    ukefFacilityId: '0030113305',
-    feeType: 'in advance',
-    feeFrequency: 'Monthly',
-    dayCountBasis: 365,
-    coverDateConfirmed: true,
-    changedToIssued: false,
-  },
-  validation: { required: [] },
-};
-
-const MOCK_REQUEST = {
-  username: 'BANK1_MAKER1',
-  firstname: 'ABC',
-  surname: 'DEF',
-  email: 'test',
-  roles: ['maker'],
-  bank: {
-    id: '9',
-    name: 'UKEF test bank (Delegated)',
-    mga: ['mga_ukef_1.docx', 'mga_ukef_2.docx'],
-    emails: [
-      'maker1@ukexportfinance.gov.uk',
-      'checker1@ukexportfinance.gov.uk',
-    ],
-    companiesHouseNo: 'UKEF0001',
-    partyUrn: '00318345',
-  },
-  timezone: 'Europe/London',
-  lastLogin: '1638791497263',
-  'user-status': 'active',
-  _id: '619bae3467cc7c002069fc1e',
-};
-
 const MOCK_DEAL = {
   _id: '61a7710b2ae62b0013dae687',
   dealType: 'GEF',
@@ -542,75 +417,6 @@ const MOCK_DEAL = {
     'user-status': 'active',
     _id: '619bae3467cc7c002069fc21',
   },
-};
-
-const MOCK_FACILITY = {
-  items: [
-    {
-      status: CONSTANTS.DEAL_STATUS.COMPLETED,
-      details: {
-        _id: '61a7714f2ae62b0013dae689',
-        dealId: '61a7710b2ae62b0013dae687',
-        type: 'CASH',
-        hasBeenIssued: true,
-        name: 'Facility one',
-        shouldCoverStartOnSubmission: false,
-        coverStartDate: '2021-12-03T00:00:00.000Z',
-        coverEndDate: '2040-01-01T00:00:00.000Z',
-        monthsOfCover: null,
-        details: [],
-        detailsOther: '',
-        currency: { id: 'GBP' },
-        value: 1000,
-        coverPercentage: 80,
-        interestPercentage: 1,
-        paymentType: 'IN_ADVANCE_MONTHLY',
-        createdAt: 1638363471661,
-        updatedAt: 1638446928711,
-        ukefExposure: 800,
-        guaranteeFee: 0.9,
-        submittedAsIssuedDate: '1638363717231',
-        ukefFacilityId: '0030113306',
-        feeType: 'in advance',
-        feeFrequency: 'Monthly',
-        dayCountBasis: 360,
-        coverDateConfirmed: null,
-      },
-      validation: { required: [] },
-    },
-    {
-      status: CONSTANTS.DEAL_STATUS.COMPLETED,
-      details: {
-        _id: '61a771cc2ae62b0013dae68a',
-        dealId: '61a7710b2ae62b0013dae687',
-        type: 'CASH',
-        hasBeenIssued: true,
-        name: 'Facility two',
-        shouldCoverStartOnSubmission: true,
-        coverStartDate: 1638403200000,
-        coverEndDate: '2030-01-01T00:00:00.000Z',
-        monthsOfCover: null,
-        details: [],
-        detailsOther: '',
-        currency: 'GBP',
-        value: 2000,
-        coverPercentage: 80,
-        interestPercentage: 1,
-        paymentType: 'IN_ADVANCE_MONTHLY',
-        createdAt: 1638363596947,
-        updatedAt: 1638442632540,
-        ukefExposure: 1600,
-        guaranteeFee: 0.9,
-        submittedAsIssuedDate: '1638363717231',
-        ukefFacilityId: '0030113305',
-        feeType: 'in advance',
-        feeFrequency: 'Monthly',
-        dayCountBasis: 365,
-        coverDateConfirmed: true,
-      },
-      validation: { required: [] },
-    },
-  ],
 };
 
 const MOCK_AIN_APPLICATION_FALSE_COMMENTS = {
@@ -721,12 +527,7 @@ const MOCK_AIN_APPLICATION_FALSE_COMMENTS = {
 module.exports = {
   MOCK_AIN_APPLICATION,
   MOCK_AIN_APPLICATION_UNISSUED_ONLY,
-  MOCK_UNISSUED_FACILITY,
-  MOCK_FACILITY,
-  MOCK_DEAL,
-  MOCK_REQUEST,
-  MOCK_ISSUED_FACILITY,
   MOCK_AIN_APPLICATION_ISSUED_ONLY,
-  MOCK_ISSUED_FACILITY_UNCHANGED,
+  MOCK_DEAL,
   MOCK_AIN_APPLICATION_FALSE_COMMENTS,
 };
