+++ gef-ui/server/utils/helpers.test.js
@@ -9,38 +9,28 @@ import {
   apiErrorHandler,
   isTrueSet,
   stringToBoolean,
-  isNotice,
-  isUkefReviewAvailable,
-  isUkefReviewPositive,
   getUTCDate,
   getEpoch,
-  pastDate,
-  futureDateInRange,
-  getFacilityCoverStartDate,
-  getIssuedFacilitiesAsArray,
-  coverDatesConfirmed,
-  makerCanReSubmit,
-  hasChangedToIssued,
-  facilitiesChangedToIssuedAsArray,
-  areUnissuedFacilitiesPresent,
   summaryItemsConditions,
-  facilityIssueDeadline,
   displayTaskComments,
 } from './helpers';
-import CONSTANTS from '../constants';
+
+import { makerCanReSubmit } from './deal-helpers';
 
 import {
   MOCK_AIN_APPLICATION,
+  MOCK_DEAL,
+  MOCK_AIN_APPLICATION_ISSUED_ONLY,
+  MOCK_AIN_APPLICATION_FALSE_COMMENTS,
+} from './MOCKS/MOCK_APPLICATIONS';
+
+import {
   MOCK_UNISSUED_FACILITY,
   MOCK_ISSUED_FACILITY,
-  MOCK_FACILITY,
-  MOCK_DEAL,
-  MOCK_REQUEST,
-  MOCK_AIN_APPLICATION_UNISSUED_ONLY,
-  MOCK_AIN_APPLICATION_ISSUED_ONLY,
   MOCK_ISSUED_FACILITY_UNCHANGED,
-  MOCK_AIN_APPLICATION_FALSE_COMMENTS,
-} from './MOCKS';
+} from './MOCKS/MOCK_FACILITIES';
+
+import { MOCK_REQUEST } from './MOCKS/MOCK_REQUESTS';
 
 describe('userToken()', () => {
   it('returns the correct user token', () => {
@@ -704,74 +694,6 @@ describe('isTrueSet()', () => {
   });
 });
 
-describe('isNotice()', () => {
-  it('Should return TRUE for any `Notice` submission type i.e. MIN or AIN', () => {
-    expect(isNotice('Manual inclusion notice')).toEqual(true);
-  });
-
-  it('Should return FALSE for any `Application` submission type i.e. MIA', () => {
-    expect(isNotice('Manual inclusion application')).toEqual(false);
-  });
-
-  it('Should return FALSE for any `Application` submission type i.e. MIA with mixed case', () => {
-    expect(isNotice('manUAL InClUsIoN APPLICATION')).toEqual(false);
-  });
-});
-
-describe('isUkefReviewAvailable()', () => {
-  it('Should return TRUE for application with UkefDecision set to `Accepted (with conditions)` and status set to `Ready for Checkers approval`', () => {
-    MOCK_DEAL.ukefDecision = [{
-      text: '1. Condition 1\r\n2. Condition 2\r\n3. Condition 3',
-      decision: 'Accepted (with conditions)',
-      timestamp: 1639657163,
-    }];
-    expect(isUkefReviewAvailable(MOCK_DEAL.status, MOCK_DEAL.ukefDecision)).toEqual(true);
-  });
-
-  it('Should return FALSE for application with no UkefDecision and status set to `Ready for Checkers approval`', () => {
-    MOCK_DEAL.ukefDecision = [];
-    expect(isUkefReviewAvailable(MOCK_DEAL.status, MOCK_DEAL.ukefDecision)).toEqual(false);
-  });
-});
-
-describe('isUkefReviewPositive()', () => {
-  it('Should return TRUE for application with UkefDecision set to `Accepted (with conditions)` and status set to `Ready for Checkers approval`', () => {
-    MOCK_DEAL.ukefDecision = [{
-      text: '1. Condition 1\r\n2. Condition 2\r\n3. Condition 3',
-      decision: 'Accepted (with conditions)',
-      timestamp: 1639657163,
-    }];
-    expect(isUkefReviewPositive(MOCK_DEAL.status, MOCK_DEAL.ukefDecision)).toEqual(true);
-  });
-
-  it('Should return FALSE for application with UkefDecision set to `Rejected by UKEF` and status set to `Ready for Checkers approval`', () => {
-    MOCK_DEAL.ukefDecision = [
-      {
-        text: '1. Condition 1\r\n2. Condition 2\r\n3. Condition 3',
-        decision: 'Rejected by UKEF',
-        timestamp: 1639657163,
-      },
-    ];
-    expect(isUkefReviewPositive(MOCK_DEAL.status, MOCK_DEAL.ukefDecision)).toEqual(false);
-  });
-
-  it('Should return FALSE for application with no UkefDecision and status set to `Ready for Checkers approval`', () => {
-    MOCK_DEAL.ukefDecision = [];
-    expect(isUkefReviewPositive(MOCK_DEAL.status, MOCK_DEAL.ukefDecision)).toEqual(false);
-  });
-
-  it('Should return TRUE for application with UkefDecision set to `Accepted (with conditions)` and status set to UKEF_APPROVED_WITHOUT_CONDITIONS status', () => {
-    const UKEF_APPPROVED_DEAL = MOCK_DEAL;
-    UKEF_APPPROVED_DEAL.status = CONSTANTS.DEAL_STATUS.UKEF_APPROVED_WITHOUT_CONDITIONS;
-    UKEF_APPPROVED_DEAL.ukefDecision = [{
-      text: '1. Condition 1\r\n2. Condition 2\r\n3. Condition 3',
-      decision: 'Accepted (with conditions)',
-      timestamp: 1639657163,
-    }];
-    expect(isUkefReviewPositive(UKEF_APPPROVED_DEAL.status, UKEF_APPPROVED_DEAL.ukefDecision)).toEqual(true);
-  });
-});
-
 describe('stringToBoolean', () => {
   it('returns `true` string as boolean', () => {
     expect(stringToBoolean('true')).toEqual(true);
@@ -796,149 +718,6 @@ describe('getEpoch', () => {
   });
   it('Should return EPOCH for the specified day, month and year custom date object', () => {
     expect(getEpoch({ year: 2021, month: 1, day: 1 })).toEqual(1609459200000);
-  });
-});
-
-describe('pastDate', () => {
-  it('Should return TRUE for the specified date', () => {
-    expect(pastDate({ day: 1, month: 1, year: 1970 })).toEqual(true);
-  });
-  it('Should return TRUE for the specified date', () => {
-    expect(pastDate({ day: 20, month: 9, year: 1989 })).toEqual(true);
-  });
-  it('Should return FALSE for the specified date', () => {
-    const date = new Date();
-    expect(pastDate({ day: date.getDate(), month: (date.getMonth() + 1), year: date.getFullYear() })).toEqual(false);
-  });
-});
-
-describe('futureDateInRange', () => {
-  it('Should return FALSE for the specified day and range days', () => {
-    expect(futureDateInRange({ day: 1, month: 1, year: 1970 }, 30)).toEqual(false);
-  });
-  it('Should return FALSE for the specified day and range days', () => {
-    expect(futureDateInRange({ day: 1, month: 1, year: 9999 }, 30)).toEqual(false);
-  });
-  it('Should return TRUE for the specified day and range days', () => {
-    const date = new Date();
-    expect(futureDateInRange({ day: date.getDate(), month: (date.getMonth() + 1), year: date.getFullYear() }, 365)).toEqual(true);
-  });
-
-  describe('getFacilityCoverStartDate', () => {
-    it('Should return expected date object for mock facility', () => {
-      const expected = {
-        date: '2',
-        month: '12',
-        year: '2021',
-      };
-      expect(getFacilityCoverStartDate(MOCK_FACILITY.items[1])).toEqual(expected);
-    });
-
-    it('Should return expected date object for mock facility', () => {
-      const expected = {
-        date: '3',
-        month: '12',
-        year: '2021',
-      };
-      expect(getFacilityCoverStartDate(MOCK_FACILITY.items[0])).toEqual(expected);
-    });
-  });
-});
-
-describe('getIssuedFacilitiesAsArray', () => {
-  it('Should return the expected facilities object from mock facilities array where the facility date has not been confirmed by the bank', () => {
-    const expected = [
-      [
-        { text: 'Facility one' },
-        { text: '0030113306' },
-        { text: 'GBP 1,000.00' },
-        { html: "<a href = '/gef/application-details/61a7710b2ae62b0013dae687/61a7714f2ae62b0013dae689/confirm-cover-start-date' class = 'govuk-button govuk-button--secondary govuk-!-margin-0'>Update</a>" },
-      ],
-    ];
-    expect(getIssuedFacilitiesAsArray(MOCK_FACILITY)).toEqual(expected);
-  });
-
-  it('Should return the empty array', () => {
-    const expected = [];
-    MOCK_FACILITY.items[0].details.hasBeenIssued = false;
-    expect(getIssuedFacilitiesAsArray(MOCK_FACILITY)).toEqual(expected);
-  });
-});
-
-describe('coverDatesConfirmed', () => {
-  it('Should return FALSE as one of the facility\'s cover date has not been confirmed', () => {
-    MOCK_FACILITY.items[0].details.hasBeenIssued = true;
-    expect(coverDatesConfirmed(MOCK_FACILITY)).toEqual(false);
-  });
-});
-
-describe('makerCanReSubmit', () => {
-  it('Should return FALSE since the deal status is BANK_CHECK', () => {
-    MOCK_DEAL.status = CONSTANTS.DEAL_STATUS.BANK_CHECK;
-    expect(makerCanReSubmit(MOCK_REQUEST, MOCK_DEAL)).toEqual(false);
-  });
-  it('Should return TRUE as the deal status has been changed to UKEF_APPROVED_WITH_CONDITIONS', () => {
-    MOCK_DEAL.status = CONSTANTS.DEAL_STATUS.UKEF_APPROVED_WITH_CONDITIONS;
-    expect(makerCanReSubmit(MOCK_REQUEST, MOCK_DEAL)).toEqual(true);
-  });
-  it('Should return TRUE as the deal has changed facilities, is AIN and has status UKEF_AKNOWLEDGED', () => {
-    expect(makerCanReSubmit(MOCK_REQUEST, MOCK_AIN_APPLICATION)).toEqual(true);
-  });
-  it('Should return FALSE as the deal does not have changed facilities, is AIN and has status UKEF_AKNOWLEDGED', () => {
-    expect(makerCanReSubmit(MOCK_REQUEST, MOCK_AIN_APPLICATION_UNISSUED_ONLY)).toEqual(false);
-  });
-});
-
-describe('hasChangedToIssued()', () => {
-  it('should return true when there are facilities changed to issued', () => {
-    expect(hasChangedToIssued(MOCK_AIN_APPLICATION)).toEqual(true);
-  });
-  it('should return false when there are no facilities changed to issued', () => {
-    expect(hasChangedToIssued(MOCK_AIN_APPLICATION_UNISSUED_ONLY)).toEqual(false);
-  });
-});
-
-describe('facilitiesChangedToIssuedAsArray()', () => {
-  it('should return an array when there are facilities changed to issued', () => {
-    const result = facilitiesChangedToIssuedAsArray(MOCK_AIN_APPLICATION);
-
-    const expected = [
-      {
-        name: MOCK_AIN_APPLICATION.facilities.items[1].details.name,
-        id: MOCK_AIN_APPLICATION.facilities.items[1].details._id,
-      },
-    ];
-    expect(result).toEqual(expected);
-  });
-  it('should return an empty array when there are no facilities changed to issued', () => {
-    const result = facilitiesChangedToIssuedAsArray(MOCK_AIN_APPLICATION_UNISSUED_ONLY);
-
-    const expected = [];
-    expect(result).toEqual(expected);
-  });
-});
-
-describe('areUnissuedFacilitiesPresent', () => {
-  it('should return false if no unissued facilities', () => {
-    expect(areUnissuedFacilitiesPresent(MOCK_AIN_APPLICATION_ISSUED_ONLY)).toEqual(false);
-  });
-  it('should return true if only unissued facilities', () => {
-    expect(areUnissuedFacilitiesPresent(MOCK_AIN_APPLICATION_UNISSUED_ONLY)).toEqual(true);
-  });
-  it('should return true if only unissued and issued facilities', () => {
-    expect(areUnissuedFacilitiesPresent(MOCK_AIN_APPLICATION)).toEqual(true);
-  });
-});
-
-describe('facilityIssueDeadline()', () => {
-  const timestamp = 1639586124100;
-
-  it('should return a correct timestamp 3 months in advance in the right format', () => {
-    const result = facilityIssueDeadline(timestamp);
-
-    const expected = '15 Mar 2022';
-
-    expect(result).toEqual(expected);
   });
 });
